requirements.txt
streamlit>=1.28.1
pandas>=2.0.3
numpy>=1.24.0
plotly>=5.15.0
firebase-admin>=6.1.0
python-dotenv>=1.0.0
reportlab>=4.0.4
python-dateutil>=2.8.2
sendgrid
pyrebase4
